Fixed type hierarchy for numeric versions
