+++ dough-common/src/main/java/io/github/bakedlibs/dough/versions/SimpleNumericVersion.java
@@ -2,20 +2,13 @@ package io.github.bakedlibs.dough.versions;
 
 import javax.annotation.Nonnull;
 
-import org.apache.commons.lang.Validate;
-
 /**
  * This {@link Version} implementation consists of only one component, the version number.
  * 
  * @author TheBusyBiscuit
  *
  */
-public class SimpleNumericVersion implements Version {
-
-    /**
-     * The version number.
-     */
-    private final int versionNumber;
+public class SimpleNumericVersion extends AbstractNumericVersion {
 
     /**
      * This constructs a new {@link SimpleNumericVersion} with the given version number.
@@ -25,9 +18,7 @@ public class SimpleNumericVersion implements Version {
      *            The version number
      */
     public SimpleNumericVersion(int version) {
-        Validate.isTrue(version > 0, "The version must be a positive number.");
-
-        this.versionNumber = version;
+        super(version);
     }
 
     /**
@@ -42,53 +33,8 @@ public class SimpleNumericVersion implements Version {
      * {@inheritDoc}
      */
     @Override
-    public boolean isNewerThan(@Nonnull Version version) {
-        if (isSimilar(version)) {
-            return getVersionNumber() > ((SimpleNumericVersion) version).getVersionNumber();
-        } else {
-            throw new IncomparableVersionsException(this, version);
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean isEqualTo(@Nonnull Version version) {
-        if (isSimilar(version)) {
-            return getVersionNumber() == ((SimpleNumericVersion) version).getVersionNumber();
-        } else {
-            throw new IncomparableVersionsException(this, version);
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean isOlderThan(@Nonnull Version version) {
-        if (isSimilar(version)) {
-            return getVersionNumber() < ((SimpleNumericVersion) version).getVersionNumber();
-        } else {
-            throw new IncomparableVersionsException(this, version);
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
     public @Nonnull String getAsString() {
-        return String.valueOf(versionNumber);
-    }
-
-    /**
-     * This method returns the version number which this numeric {@link Version} represents.
-     * 
-     * @return The version number
-     */
-    public final int getVersionNumber() {
-        return versionNumber;
+        return String.valueOf(getVersionNumber());
     }
 
     /**
@@ -96,7 +42,7 @@ public class SimpleNumericVersion implements Version {
      */
     @Override
     public int hashCode() {
-        return versionNumber;
+        return getVersionNumber();
     }
 
     /**
